Inspector edit + add completed.
